moving d3 code to separate file
